feat: fix add form structure
